# WordBattleGame/Docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WordBattleGame",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5098"
    }
  ],
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerRegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerRegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerLoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerLoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/update-profile/{id}": {
      "put": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/change-password/{id}": {
      "put": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Game/generate-word": {
      "post": {
        "tags": [
          "Game"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWordRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWordRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedWordResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedWordResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedWordResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/create-round": {
      "post": {
        "tags": [
          "Game"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players": {
      "get": {
        "tags": [
          "Players"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}": {
      "get": {
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateRoundRequestDto": {
        "required": [
          "gameId"
        ],
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Game": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "maxRound": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "nullable": true
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Round"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GenerateWordRequestDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeneratedWordResponseDto": {
        "type": "object",
        "properties": {
          "generatedWord": {
            "type": "string",
            "nullable": true
          },
          "trueWord": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "player": {
            "$ref": "#/components/schemas/PlayerDto"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Player": {
        "required": [
          "email",
          "id",
          "name",
          "passwordHash"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "stats": {
            "$ref": "#/components/schemas/PlayerStats"
          }
        },
        "additionalProperties": false
      },
      "PlayerDto": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PlayerLoginDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerRegisterDto": {
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerStats": {
        "required": [
          "id",
          "playerId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "playerId": {
            "type": "string",
            "nullable": true
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          },
          "totalGames": {
            "type": "integer",
            "format": "int32"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "win": {
            "type": "integer",
            "format": "int32"
          },
          "lose": {
            "type": "integer",
            "format": "int32"
          },
          "draw": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Round": {
        "required": [
          "gameId",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "gameId": {
            "type": "string",
            "nullable": true
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "generatedWord": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "nullable": true
          },
          "trueWord": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "winnerId": {
            "type": "string",
            "nullable": true
          },
          "winner": {
            "$ref": "#/components/schemas/Player"
          }
        },
        "additionalProperties": false
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}